+++ imports/experiment/server/index.js
@@ -1,6 +1,6 @@
 import { conditions } from "./game/conditions.js";
 import { init } from "./game/init.js";
-import { onRoundEnd, onRoundStart, onStageEnd } from "./game/callbacks.js";
+import { onRoundEnd, onRoundStart, onStageEnd, onGameEnd} from "./game/callbacks.js";
 
 export const config = {
   conditions,
@@ -8,5 +8,6 @@ export const config = {
   onRoundStart,
   onStageEnd,
   onRoundEnd,
+  onGameEnd,
   bots: []
 };